renamed mtime_cached_property to contents_property
