+++ .pre-commit-config.yaml
@@ -1,39 +1,45 @@
 # See https://pre-commit.com for more information
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
-    - repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v4.3.0
-      hooks:
-          - id: trailing-whitespace
-          - id: end-of-file-fixer
-          - id: check-yaml
-          - id: check-added-large-files
-    - repo: https://github.com/psf/black
-      rev: 22.3.0
-      hooks:
-          - id: black
-            exclude: ^(fileformats/core/_version\.py)$
-            args:
-                - -l 88
-    - repo: https://github.com/codespell-project/codespell
-      rev: v2.1.0
-      hooks:
-          - id: codespell
-            exclude: ^(fileformats/core/_version\.py)$
-            args:
-                - --ignore-words=.codespell-ignorewords
-    - repo: https://github.com/PyCQA/flake8
-      rev: 4.0.1
-      hooks:
-          - id: flake8
-    - repo: https://github.com/pre-commit/mirrors-mypy
-      rev: v1.11.2
-      hooks:
-          - id: mypy
-            args:
-                [
-                    --strict,
-                    --install-types,
-                    --non-interactive,
-                    --no-warn-unused-ignores,
-                ]
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.3.0
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-yaml
+      - id: check-added-large-files
+  - repo: https://github.com/psf/black
+    rev: 22.3.0
+    hooks:
+      - id: black
+        exclude: ^(fileformats/core/_version\.py)$
+        args:
+          - -l 88
+  - repo: https://github.com/codespell-project/codespell
+    rev: v2.1.0
+    hooks:
+      - id: codespell
+        exclude: ^(fileformats/core/_version\.py)$
+        args:
+          - --ignore-words=.codespell-ignorewords
+  - repo: https://github.com/PyCQA/flake8
+    rev: 4.0.1
+    hooks:
+      - id: flake8
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v1.11.2
+    hooks:
+      - id: mypy
+        args:
+          [
+            --strict,
+            --install-types,
+            --non-interactive,
+            --no-warn-unused-ignores,
+            --exclude,
+            tests,
+            --exclude,
+            scripts,
+            --exclude,
+            build,
+          ]
