+++ fileformats/core/utils.py
@@ -1,12 +1,10 @@
 import importlib
 from pathlib import Path
 import inspect
-import sys
 import typing as ty
 from types import ModuleType
 import urllib.request
 import urllib.error
-from threading import RLock
 import os
 import logging
 import pkgutil
@@ -94,25 +92,6 @@ def fspaths_converter(fspaths: FspathsInputType) -> ty.FrozenSet[Path]:
     elif isinstance(fspaths, (str, os.PathLike)):
         fspaths = [Path(fspaths)]
     return frozenset(Path(p).absolute() for p in fspaths)
-
-
-PropReturn = ty.TypeVar("PropReturn")
-
-
-def classproperty(meth: ty.Callable[..., PropReturn]) -> PropReturn:
-    """Access a @classmethod like a @property."""
-    # mypy doesn't understand class properties yet: https://github.com/python/mypy/issues/2563
-    return classmethod(property(meth))  # type: ignore
-
-
-if sys.version_info[:2] < (3, 9):
-
-    class classproperty(object):  # type: ignore[no-redef]  # noqa
-        def __init__(self, f: ty.Callable[[ty.Type[ty.Any]], ty.Any]):
-            self.f = f
-
-        def __get__(self, obj: ty.Any, owner: ty.Any) -> ty.Any:
-            return self.f(owner)
 
 
 def add_exc_note(e: Exception, note: str) -> Exception:
@@ -249,54 +228,3 @@ def import_extras_module(klass: ty.Type["fileformats.core.DataType"]) -> ExtrasM
     else:
         extras_imported = True
     return ExtrasModule(extras_imported, extras_pkg, extras_pypi)
-
-
-ReturnType = ty.TypeVar("ReturnType")
-
-
-class mtime_cached_property:
-    """A property that is cached until the mtimes of the files in the fileset are changed"""
-
-    def __init__(self, func: ty.Callable[..., ty.Any]):
-        self.func = func
-        self.__doc__ = func.__doc__
-        self.lock = RLock()
-
-    @property
-    def _cache_name(self) -> str:
-        return f"_{self.func.__name__}_mtime_cache"
-
-    def clear(self, instance: "fileformats.core.FileSet") -> None:
-        """Forcibly clear the cache"""
-        del instance.__dict__[self._cache_name]
-
-    def __get__(
-        self,
-        instance: ty.Optional["fileformats.core.FileSet"],
-        owner: ty.Optional[ty.Type["fileformats.core.FileSet"]] = None,
-    ) -> ty.Any:
-        if instance is None:  # if accessing property from class not instance
-            return self
-        assert isinstance(instance, fileformats.core.FileSet), (
-            "Cannot use mtime_cached_property instance with "
-            f"{type(instance).__name__!r} object, only FileSet objects."
-        )
-        try:
-            mtimes, value = instance.__dict__[self._cache_name]
-        except KeyError:
-            pass
-        else:
-            if instance.mtimes == mtimes:
-                return value
-        with self.lock:
-            # check if another thread filled cache while we awaited lock
-            try:
-                mtimes, value = instance.__dict__[self._cache_name]
-            except KeyError:
-                pass
-            else:
-                if instance.mtimes == mtimes:
-                    return value
-            value = self.func(instance)
-            instance.__dict__[self._cache_name] = (instance.mtimes, value)
-        return value
